update hiers algo, move to separate file
