+++ app/requirements.txt
@@ -5,6 +5,7 @@ python-dotenv
 flasgger
 rdkit
 scaffoldgraph[rdkit]
+useful_rdkit_utils
 pandas
 psycopg2
 gunicorn
